# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/provider": {
      "get": {
        "operationId": "ProviderController_getAllProviders",
        "parameters": [
          {
            "name": "search-term",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profession",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "town",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "specialization",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "experience",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Providers"
        ]
      },
      "post": {
        "operationId": "ProviderController_createProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
    "/provider/{id}": {
      "get": {
        "operationId": "ProviderController_getProviderById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          }
        },
        "tags": [
          "Providers"
        ]
      },
      "delete": {
        "operationId": "ProviderController_deleteProvider",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
    "/provider/me": {
      "get": {
        "operationId": "ProviderController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          }
        },
        "tags": [
          "Providers"
        ]
      },
      "patch": {
        "operationId": "ProviderController_updateProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
...
    "/provider/{id}/photo": {
      "delete": {
        "operationId": "ProviderController_deleteProviderPhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "Providers"
        ]
      }
    },
...
    "/user/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/customer": {
      "get": {
        "operationId": "CustomerController_getAllCustomers",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "CustomerController_createCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
    "/customer/me": {
      "get": {
        "operationId": "CustomerController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomerController_updateCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        },
        "tags": [
          "Customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Labour Connect API",
    "description": "A RESTful API for Labour Connect platform built with NestJS, Knex.js, and TypeScript",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "facebookLink": {
            "type": "string"
          },
          "twitterLink": {
            "type": "string"
          },
          "instagramLink": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "experience": {
            "type": "string"
          },
          "specializations": {
            "type": "string"
          },
          "qualifications": {
            "type": "string"
          },
          "availability": {
            "type": "string"
          },
          "availableDays": {
            "type": "string"
          },
          "completedJobs": {
            "type": "number"
          },
          "verified": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended",
              "pending"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "createdDate": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "profession",
          "address",
          "town",
          "city",
          "country",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProviderResponse": {
        "type": "object",
        "properties": {
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Provider"
            }
          }
        },
        "required": [
          "providers"
        ]
      },
      "Buffer": {
        "type": "object",
        "properties": {}
      },
      "AttachmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "buffer": {
            "$ref": "#/components/schemas/Buffer"
          }
        },
        "required": [
          "name",
          "size",
          "type",
          "data"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "qualifications": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "facebookLink": {
            "type": "string"
          },
          "twitterLink": {
            "type": "string"
          },
          "instagramLink": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          },
          "experience": {
            "type": "number"
          },
          "specializations": {
            "type": "string"
          },
          "availability": {
            "type": "string"
          },
          "availableDays": {
            "type": "string"
          },
          "completedJobs": {
            "type": "number"
          },
          "verified": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended",
              "pending"
            ]
          },
          "photo": {
            "$ref": "#/components/schemas/AttachmentDto"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttachmentDto"
            }
          }
        }
      },
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "qualifications": {
            "type": "string"
          },
          "facebookLink": {
            "type": "string"
          },
          "twitterLink": {
            "type": "string"
          },
          "instagramLink": {
            "type": "string"
          },
          "experience": {
            "type": "string"
          },
          "specializations": {
            "type": "string"
          },
          "availability": {
            "type": "string"
          },
          "availableDays": {
            "type": "string"
          },
          "completedJobs": {
            "type": "number"
          },
          "createdDate": {
            "type": "string"
          },
          "photo": {
            "$ref": "#/components/schemas/AttachmentDto"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "phone",
          "address",
          "town",
          "city",
          "country"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "auth0Id": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "role": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "providerId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "auth0Id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "suspended"
            ]
          },
          "registrationDate": {
            "type": "string"
          },
          "totalRequests": {
            "type": "number"
          },
          "rating": {
            "type": "number"
          },
          "verified": {
            "type": "boolean"
          },
          "serviceKeywords": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "address",
          "status",
          "registrationDate",
          "totalRequests",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "customers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          }
        },
        "required": [
          "customers"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "serviceKeywords": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "phone",
          "address"
        ]
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "serviceKeywords": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
...
    }
  }
}